Provide a temporary fix for commit 4b1f8ce60d21cb6f404215bdd1ea335e26a6b869
If a package version changes in the repo, the call to apt-get -t $release fails
because of this. This causes the re-installation of all packages from that
specific release to fail. This in turns causes the cleanup function to fail.

A temporary fix is to let APT install the lastest version from the repo.